dataplacement: update ruh info initialization
The current way of initilization limits ruhs from 0 to 127.
Although fio supports a maximum of 128 ruhs, for the files supporting
more the 128 placement ids user has the flexibility of specifying any
128 placement id index at a time. This commit updates the way we fetch
ruhs. We now fetch the ruhs info in two steps. The first step only gets
us the number of ruhs from the ioengine. This is used by fdp backend to
allocate the correct buffer size for the second step, where we fetch
the actual ruhs info.

Signed-off-by: Ankit Kumar <ankit.kumar@samsung.com>