Fixes issue where Ki,Mi,Gi,Ti,Pi behave like K,KB etc instead of KiB.
As described in issue 1618, the Ki,Mi,etc units do not behave as expected.
The code in parse.c explicitly handles K,KB and KiB, but not Ki,
this results in Ki being implicitly handled by the "K" check.
This commit adds explicit handling of Ki,Mi, etc to match the behavior of
KiB, MiB, etc.
This change also brings the behaviour in line with the documentation,
though the examples have incorrect numbers for TB and have duplicated
entries between kb_base=1000 and kb_base=1024.

Sean Ashton <sean@ashtech.one>