Improve flushing pagecache on darwin
The current behavior of flushing the pagecache via
posix_fadvise(POSIX_FADV_DONTNEED) is not a valid operation on macOS.
This can result adverse outcomes when measuring performance.
The change will instead flush via mmap and msync, which better reflects
the desired behavior on macOS.

The outcome is cached reads should no longer manifest as a gain.

- Move to __APPLE__
- Migrate code to os/os-mac.h since it is about macOS

Signed-off-by: DeveloperEcosystemEngineering@apple.com