ci: don't skip verify tests when triggered manually
When our test suite is triggered manually it skips the verify test
script. Change this so that the verify test script is run when the tests
are manually kicked off. This is accomplished by skipping the verify
test suite only when a push or pull request triggers our tests.

With this change we don't have to wait overnight to see the results of
the verify tests.

Link: https://lore.kernel.org/r/20250508185832.3702-8-vincent.fu@samsung.com
Signed-off-by: Vincent Fu <vincent.fu@samsung.com>