backend: get io_u from requeue if exists
If the `iodepth=1`, which means only 1 io_u can alive, and also if it is
in `td->io_u_requeues`, the `td->io_u_freelist` is empty. With
`end_fsync=1` in this, `__get_io_u()` will return `NULL`, because there
are no free io_u in `td->io_u_freelist` and `td->runstate` is
`FSYNCING`. This leads that the sync operation will be failed with
message `end_fsync failed for file <filename>`.

This patch fixes that the io_u's in `td->io_u_requeues` will be returned
if `td->runstate == TD_FSYNCING`, and will not be set the
`IO_U_F_NO_FILE_PUT` flag, which causes the file's reference count never
goes to 0.

Signed-off-by: Jonghwi Jeong <jongh2.jeong@samsung.com>