REPORTING-BUGS: add note about reports on malicious job files
Congratulations, you managed to construct an odd job file that caused
the fio parser to crash. Such as the following examples:

https://github.com/axboe/fio/issues/1981
https://github.com/axboe/fio/issues/1982
https://github.com/axboe/fio/issues/1994

When the answer is literally "well don't do that then", don't bother
reporting the issue unless YOU intend to fix it yourself. It's not
a priority item for the fio project.

Signed-off-by: Jens Axboe <axboe@kernel.dk>