io_u: put io_u->file from requeue in FSYNCING
When iodepth=1 and ->queue() returns FIO_Q_BUSY, the only io_u is pushed
to io_u_requeues. With end_fsync=1, the backend cannot retrieve io_u
from either io_u_freelist or io_u_requeues, leading to end_fsync
failure.

This patch fixes the issue by obtaining an io_u from the requeue list
even when fio is fsyncing.  However, when fio is fsyncing, this patch
makes sure to decrement the file reference counter for the file
associated with the io_u since it is possible that the fsync will target
a different file.  The larger context of this patch is to ensure that
file references are correctly counted so that fio can close all files on
exit.

Fixes: 1c83ed1 ("io_u: get io_u from io_u_freelist when TD_FSYNCING")
Signed-off-by: Jonghwi Jeong <jongh2.jeong@samsung.com>