engines/mmap: fix logic when "offset" > 0
This fixes the offset/size calculation when "offset" > 0. Previously this had:

        fmd->mmap_sz = f->io_size;
        fmd->mmap_off = 0;

This is correct only when "offset" (ie f->file_offset) = 0. We can fix this by
either keeping the offset at 0, and map the entire file, or by keeping the size
at f->io_size and setting mmap_off = f->file_offset. On 64-bit there isn't much
difference but on 32-bit the second option is better as it maps less memory.
This patch therefore takes the second approach.